# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gaming Lounge OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── TENANTS ──────────────────────────────────────────
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets for current tenant
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{assetId}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get an asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{assetId}/qr:
    post:
      operationId: generateAssetQr
      tags: [assets]
      summary: Generate QR code for an asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"

  # ── SESSIONS ──────────────────────────────────────────
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts
      responses:
        "200":
          description: Shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: openShift
      tags: [shifts]
      summary: Open a new shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftInput"
      responses:
        "201":
          description: Shift opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
...
  /dashboard/revenue:
    get:
      operationId: getRevenueStats
      tags: [dashboard]
      summary: Get revenue statistics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month]
      responses:
        "200":
          description: Revenue stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueStats"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── TENANT SCHEMAS ──
...
    User:
      type: object
      required: [id, email, name, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        role:
          type: string
          enum: [platform_owner, owner, manager, cashier, buffet_worker]
        tenantId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Asset:
      type: object
      required: [id, name, type, pricePerHour, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        type:
          type: string
          enum: [ps, billiard, air_hockey, babyfoot, other]
        pricePerHour:
          type: number
        status:
          type: string
          enum: [available, busy]
        qrToken:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AssetInput:
      type: object
      required: [name, type, pricePerHour]
      properties:
        name:
          type: string
        nameAr:
          type: string
        type:
          type: string
          enum: [ps, billiard, air_hockey, babyfoot, other]
        pricePerHour:
          type: number
    AssetUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        type:
          type: string
          enum: [ps, billiard, air_hockey, babyfoot, other]
        pricePerHour:
          type: number
        status:
          type: string
          enum: [available, busy]
    QrCode:
      type: object
      required: [token, assetId]
      properties:
        token:
          type: string
        assetId:
          type: integer
        qrDataUrl:
          type: ["string", "null"]

    # ── SESSION SCHEMAS ──
...
    Shift:
      type: object
      required: [id, userId, status, openingCash, openedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, closed]
        openingCash:
          type: number
        expectedCash:
          type: ["number", "null"]
        actualCash:
          type: ["number", "null"]
        difference:
          type: ["number", "null"]
        differenceExplanation:
          type: ["string", "null"]
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
    ShiftInput:
      type: object
      required: [openingCash]
      properties:
        openingCash:
          type: number
...
    RevenueStats:
      type: object
      required: [total, sessionRevenue, orderRevenue, period]
      properties:
        total:
          type: number
        sessionRevenue:
          type: number
        orderRevenue:
          type: number
        period:
          type: string
        paymentMethodBreakdown:
          type: object
          properties:
            cash:
              type: number
            instapay:
              type: number
            visa:
              type: number
...